# backend/openapi.json
{
  "components": {
    "schemas": {
      "DeviceResponse": {
        "additionalProperties": false,
        "description": "Device capabilities with the shared API response envelope fields.",
        "properties": {
          "apiVersion": {
            "const": "1",
            "description": "Expose the immutable contract version as a required response field.",
            "readOnly": true,
            "title": "Apiversion",
            "type": "string"
          },
          "coremlAvailable": {
            "default": false,
            "title": "Coremlavailable",
            "type": "boolean"
          },
          "cudaAvailable": {
            "title": "Cudaavailable",
            "type": "boolean"
          },
          "cudaDeviceCount": {
            "title": "Cudadevicecount",
            "type": "integer"
          },
          "deviceName": {
            "title": "Devicename",
            "type": "string"
          },
          "directmlAvailable": {
            "default": false,
            "title": "Directmlavailable",
            "type": "boolean"
          },
          "mpsAvailable": {
            "title": "Mpsavailable",
            "type": "boolean"
          },
          "onnxCudaAvailable": {
            "default": false,
            "title": "Onnxcudaavailable",
            "type": "boolean"
          },
          "onnxRuntimeAvailable": {
            "default": false,
            "title": "Onnxruntimeavailable",
            "type": "boolean"
          },
          "requestId": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
            "title": "Requestid",
            "type": "string"
          },
          "selectedDevice": {
            "enum": [
              "cpu",
              "cuda",
              "mps",
              "coreml",
              "directml"
            ],
            "title": "Selecteddevice",
            "type": "string"
          },
          "torchAvailable": {
            "title": "Torchavailable",
            "type": "boolean"
          },
          "torchCudaAvailable": {
            "default": false,
            "title": "Torchcudaavailable",
            "type": "boolean"
          },
          "totalMemoryMb": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Totalmemorymb"
          }
        },
        "required": [
          "selectedDevice",
          "torchAvailable",
          "onnxRuntimeAvailable",
          "cudaAvailable",
          "torchCudaAvailable",
          "onnxCudaAvailable",
          "mpsAvailable",
          "coremlAvailable",
          "directmlAvailable",
          "deviceName",
          "cudaDeviceCount",
          "totalMemoryMb",
          "requestId",
          "apiVersion"
        ],
        "title": "DeviceResponse",
        "type": "object"
      },
      "ErrorInfo": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "AUTHENTICATION_REQUIRED",
              "INVALID_AUTHENTICATION_TOKEN",
              "VALIDATION_ERROR",
              "CONFLICT",
              "SERVICE_UNAVAILABLE",
              "NOT_FOUND",
              "METHOD_NOT_ALLOWED",
              "HTTP_ERROR",
              "INTERNAL_ERROR"
            ],
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorInfo",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "apiVersion": {
            "const": "1",
            "description": "Expose the immutable contract version as a required response field.",
            "readOnly": true,
            "title": "Apiversion",
            "type": "string"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorInfo"
          },
          "requestId": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
            "title": "Requestid",
            "type": "string"
          },
          "success": {
            "const": false,
            "readOnly": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "requestId",
          "error",
          "apiVersion",
          "success"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "ErrorResponseV2": {
        "additionalProperties": false,
        "properties": {
          "apiVersion": {
            "const": "2",
            "description": "Expose the harmony-preview contract version.",
            "readOnly": true,
            "title": "Apiversion",
            "type": "string"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorInfo"
          },
          "requestId": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
            "title": "Requestid",
            "type": "string"
          },
          "success": {
            "const": false,
            "readOnly": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "requestId",
          "error",
          "apiVersion",
          "success"
        ],
        "title": "ErrorResponseV2",
        "type": "object"
      },
      "HarmonyCancelRequest": {
        "additionalProperties": false,
        "properties": {
          "apiVersion": {
            "const": "2",
            "title": "Apiversion",
            "type": "string"
          },
          "requestId": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
            "title": "Requestid",
            "type": "string"
          }
        },
        "required": [
          "apiVersion",
          "requestId"
        ],
        "title": "HarmonyCancelRequest",
        "type": "object"
      },
...
      "HarmonyGenerateRequest": {
        "additionalProperties": false,
        "properties": {
          "allowCpuFallback": {
            "default": true,
            "title": "Allowcpufallback",
            "type": "boolean"
          },
          "apiVersion": {
            "const": "2",
            "title": "Apiversion",
            "type": "string"
          },
          "candidateCount": {
            "default": 32,
            "maximum": 32.0,
            "minimum": 1.0,
            "title": "Candidatecount",
            "type": "integer"
          },
          "controls": {
            "$ref": "#/components/schemas/HarmonyControls"
          },
          "existingHarmony": {
            "items": {
              "$ref": "#/components/schemas/HarmonyCondition"
            },
            "maxItems": 4096,
            "title": "Existingharmony",
            "type": "array"
          },
          "generationMask": {
            "items": {
              "$ref": "#/components/schemas/HarmonyEditSpan"
            },
            "maxItems": 512,
            "minItems": 1,
            "title": "Generationmask",
            "type": "array"
          },
          "melody": {
            "items": {
              "$ref": "#/components/schemas/HarmonyMelodyNote"
            },
            "maxItems": 16384,
            "title": "Melody",
            "type": "array"
          },
          "modelId": {
            "default": "harmonyforge-bimask-base-v1",
            "enum": [
              "harmonyforge-bimask-base-v1",
              "mock-harmonyforge-bimask-v1"
            ],
            "title": "Modelid",
            "type": "string"
          },
          "preferredDevice": {
            "default": "auto",
            "enum": [
              "auto",
              "cpu",
              "cuda",
              "mps"
            ],
            "title": "Preferreddevice",
            "type": "string"
          },
          "requestId": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
            "title": "Requestid",
            "type": "string"
          },
          "seed": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Seed",
            "type": "string"
          },
          "tonalities": {
            "items": {
              "$ref": "#/components/schemas/HarmonyTonalitySpan"
            },
            "maxItems": 256,
            "minItems": 1,
            "title": "Tonalities",
            "type": "array"
          }
        },
        "required": [
          "apiVersion",
          "requestId",
          "seed",
          "melody",
          "generationMask",
          "tonalities",
          "controls"
        ],
        "title": "HarmonyGenerateRequest",
        "type": "object"
      },
...
      "HarmonyJobResponse": {
        "additionalProperties": false,
        "properties": {
          "apiVersion": {
            "const": "2",
            "description": "Expose the harmony-preview contract version.",
            "readOnly": true,
            "title": "Apiversion",
            "type": "string"
          },
          "backend": {
            "enum": [
              "pytorch",
              "mock"
            ],
            "title": "Backend",
            "type": "string"
          },
          "batchSize": {
            "default": 1,
            "maximum": 32.0,
            "minimum": 1.0,
            "title": "Batchsize",
            "type": "integer"
          },
          "candidateCount": {
            "maximum": 32.0,
            "minimum": 1.0,
            "title": "Candidatecount",
            "type": "integer"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/HarmonyCandidate"
            },
            "title": "Candidates",
            "type": "array"
          },
          "checkpointSha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkpointsha256"
          },
          "cpuFallbackUsed": {
            "default": false,
            "title": "Cpufallbackused",
            "type": "boolean"
          },
          "deterministic": {
            "default": true,
            "title": "Deterministic",
            "type": "boolean"
          },
          "device": {
            "anyOf": [
              {
                "enum": [
                  "cpu",
                  "cuda",
                  "mps",
                  "coreml",
                  "directml"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device"
          },
          "dtype": {
            "anyOf": [
              {
                "enum": [
                  "float32",
                  "float16",
                  "bfloat16"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dtype"
          },
          "elapsedMs": {
            "minimum": 0.0,
            "title": "Elapsedms",
            "type": "integer"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HarmonyJobError"
              },
              {
                "type": "null"
              }
            ]
          },
          "fallbackReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallbackreason"
          },
          "mock": {
            "title": "Mock",
            "type": "boolean"
          },
          "modelId": {
            "enum": [
              "harmonyforge-bimask-base-v1",
              "mock-harmonyforge-bimask-v1"
            ],
            "title": "Modelid",
            "type": "string"
          },
          "partialCandidateStored": {
            "const": false,
            "default": false,
            "title": "Partialcandidatestored",
            "type": "boolean"
          },
          "progress": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress",
            "type": "integer"
          },
          "requestId": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
            "title": "Requestid",
            "type": "string"
          },
          "sourceCommit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcecommit"
          },
          "stage": {
            "enum": [
              "Queued",
              "Loading checkpoint",
              "Encoding",
              "Neural proposal",
              "Decoding",
              "Schema validation",
              "Client theory validation",
              "Complete",
              "Cancel requested",
              "Cancelled",
              "Failed"
            ],
            "title": "Stage",
            "type": "string"
          },
          "stageTimingsMs": {
            "additionalProperties": {
              "minimum": 0.0,
              "type": "integer"
            },
            "title": "Stagetimingsms",
            "type": "object"
          },
          "state": {
            "enum": [
              "queued",
              "running",
              "cancelRequested",
              "completed",
              "cancelled",
              "failed"
            ],
            "title": "State",
            "type": "string"
          },
          "tokenizerSha256": {
            "title": "Tokenizersha256",
            "type": "string"
          },
          "trained": {
            "title": "Trained",
            "type": "boolean"
          }
        },
        "required": [
          "requestId",
          "state",
          "stage",
          "progress",
          "elapsedMs",
          "modelId",
          "device",
          "backend",
          "dtype",
          "mock",
          "trained",
          "tokenizerSha256",
          "candidateCount",
          "apiVersion"
        ],
        "title": "HarmonyJobResponse",
        "type": "object"
      },
...
      "ModelsResponse": {
        "additionalProperties": false,
        "properties": {
          "activeBackend": {
            "enum": [
              "linear",
              "corpus",
              "pytorch",
              "onnx",
              "browser",
              "mock"
            ],
            "title": "Activebackend",
            "type": "string"
          },
          "activeModel": {
            "enum": [
              "local-deterministic-v1",
              "harmony-corpus-ngram-v1",
              "local-mlp-v1",
              "local-onnx-v1",
              "mock-deterministic-v1"
            ],
            "title": "Activemodel",
            "type": "string"
          },
          "activeRuntime": {
            "enum": [
              "cpu",
              "cuda",
              "mps",
              "coreml",
              "directml"
            ],
            "title": "Activeruntime",
            "type": "string"
          },
          "apiVersion": {
            "const": "1",
            "description": "Expose the immutable contract version as a required response field.",
            "readOnly": true,
            "title": "Apiversion",
            "type": "string"
          },
          "fallbackReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallbackreason"
          },
          "mock": {
            "title": "Mock",
            "type": "boolean"
          },
          "models": {
            "items": {
              "$ref": "#/components/schemas/ModelInfo"
            },
            "title": "Models",
            "type": "array"
          },
          "requestId": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
            "title": "Requestid",
            "type": "string"
          }
        },
        "required": [
          "requestId",
          "models",
          "activeModel",
          "activeRuntime",
          "activeBackend",
          "mock",
          "fallbackReason",
          "apiVersion"
        ],
        "title": "ModelsResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "MtcSharedToken": {
        "description": "Shared LAN token. Required only when GET /api/health reports authRequired=true.",
        "in": "header",
        "name": "X-MTC-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Visual studio chord API",
    "version": "0.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/device": {
      "get": {
        "operationId": "getDevice",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "An internal server error occurred."
          }
        },
        "summary": "Device",
        "tags": [
          "system"
        ]
      }
    },
...
    "/api/models": {
      "get": {
        "operationId": "listModels",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "An internal server error occurred."
          }
        },
        "summary": "Models",
        "tags": [
          "system"
        ]
      }
    },
...
    "/api/v2/harmony/cancel/{request_id}": {
      "post": {
        "operationId": "cancelHarmonyGeneration",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HarmonyCancelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HarmonyJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseV2"
                }
              }
            },
            "description": "Authentication is required."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseV2"
                }
              }
            },
            "description": "Authentication token is invalid."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseV2"
                }
              }
            },
            "description": "Resource not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseV2"
                }
              }
            },
            "description": "The request conflicts with the current server state."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseV2"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseV2"
                }
              }
            },
            "description": "An internal server error occurred."
          }
        },
        "security": [
          {
            "MtcSharedToken": []
          }
        ],
        "summary": "Cancel Harmony",
        "tags": [
          "neural-harmony"
        ]
      }
    },
    "/api/v2/harmony/generate": {
      "post": {
        "operationId": "startHarmonyGeneration",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HarmonyGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HarmonyJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseV2"
                }
              }
            },
            "description": "Authentication is required."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseV2"
                }
              }
            },
            "description": "Authentication token is invalid."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseV2"
                }
              }
            },
            "description": "The request conflicts with the current server state."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseV2"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseV2"
                }
              }
            },
            "description": "An internal server error occurred."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseV2"
                }
              }
            },
            "description": "The requested service is unavailable."
          }
        },
        "security": [
          {
            "MtcSharedToken": []
          }
        ],
        "summary": "Generate Harmony",
        "tags": [
          "neural-harmony"
        ]
      }
    },
    "/api/v2/jobs/{request_id}": {
      "get": {
        "operationId": "getHarmonyJob",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HarmonyJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseV2"
                }
              }
            },
            "description": "Authentication is required."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseV2"
                }
              }
            },
            "description": "Authentication token is invalid."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseV2"
                }
              }
            },
            "description": "Resource not found."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseV2"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseV2"
                }
              }
            },
            "description": "An internal server error occurred."
          }
        },
        "security": [
          {
            "MtcSharedToken": []
          }
        ],
        "summary": "Get Harmony Job",
        "tags": [
          "neural-harmony"
        ]
      }
    },
    "/api/v2/models/{model_id}/manifest": {
      "get": {
        "operationId": "getHarmonyModelManifest",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "enum": [
                "harmonyforge-bimask-base-v1",
                "mock-harmonyforge-bimask-v1"
              ],
              "title": "Model Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HarmonyModelManifestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseV2"
                }
              }
            },
            "description": "Resource not found."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseV2"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseV2"
                }
              }
            },
            "description": "An internal server error occurred."
          }
        },
        "summary": "Harmony Manifest",
        "tags": [
          "neural-harmony"
        ]
      }
    }
  }
}